gasLimit为long，序列化不能为字符串

+++ filecoin4j-rpc/src/main/java/com/github/mufanh/filecoin4j/domain/types/Message.java
@@ -25,7 +25,7 @@ public class Message implements Serializable {
 
     private BigInteger value;
 
-    private BigInteger gasLimit;
+    private Long gasLimit;
 
     private BigInteger gasFeeCap;
